Allow controller#render to pass blocks to renderables
Currently when controllers render Renderables (objects that respond to
`#render_in`) blocks aren't passed to the renderable object. This
differs from the `render` method in the view which does allow blocks to
be passed.

This updates the rendering code to pass a block throughout the call
chain until it can be added to options, which are then used in the
rendering logic to pass the block to the `Renderable` template class so
that it can be used when `render` is called.

This change allows controllers to pass blocks to renderables, matching
the behavior of the view `render` method. It does so by passing the
block though the call chain until it can be stored in the `options`
hash. From there we pass the block to
`ActionView::Template::Renderable.new` and the block is properly passed
to the `render_in` method when `ActionView::Template::Renderable#render`
is called.

* Updates controller code to pass the block from `#render` until
  arguments are normalized.
* Stores `block` in the `options` hash that is passed throughout the
  rendering call stack.
* Updates `ActionView::Template::Renderable` to accept a block, which is
  then passed to the renderables `render_in` method.
* Adds a test validating that blocks can be passed from controller
  renderers.