Fix CollectionProxy#load_async
Fix: https://github.com/rails/rails/issues/45408

Since `CollectionProxy` inherits from `Relation`, `load_async`
would schedule the query and store in it an instance variable
but would never use it.

Ideally we should make the `Association` classes async aware so that
that you can indeed schedule an association to be loaded, but that's
quite a big refactoring, and would be hard to backport to 7-0-stable.

So this is a more focused fix. We delegate `load_async` to `scope`
which means we return an async `Relation` that can later be used.

However the association isn't marked as loaded, so if you use the
association later, you'll trigger a second query anyway.

I'll try to address this in a followup that won't be backported.