Require ruby-vips before downloading blob
This ensures that if `download_blob_to_tempfile` raises an error before
evaluating its block, `rescue ::Vips::Error` will not cause a
`NameError` due to `ruby-vips` not being loaded yet.

This also prevents the blob from being downloaded if `ruby-vips` is not
available to analyze it.

A similar change was made for `Analyzer::ImageAnalyzer::ImageMagick`
in #45420.