Support ParameterFiltering based on allow lists
`ActiveSupport::ParameterFilter` currently works as a deny list: you
specify the parameters that should be filtered. If you want to specify
the parameters that should be allowed you could do it with a lambda, for
example:

```ruby
Rails.application.config.filter_parameters += [
  lambda { |key, value|
    value.replace('[FILTERED]') unless key.in?(:id)
  }
]
```

But this can be error prone. For example this doesn't work if a value is
nil.

This change makes it possible to list the parameters that are allowed.
All other parameters will be masked.

```ruby
Rails.application.config.allow_parameters = [:id]
```

`allow_parameters` takes precedence over `filter_parameters`. So if the
`allow_parameters` are defined `filter_parameters` will be ignored. So
filtered parameters

The current implementation of ParameterAllowFilter supports all features
from ParameterFilter, including lambdas. But maybe it should be
simplified instead by removing lambda support.