Fix adding sprockets-rails during upgrade
The previous strategy for adding `sprockets-rails` did not work for two
reasons:
- the regular expression would only match `require("rails/all")` (with
  parentheses) and not `require "rails/all"`
- adding a require in `config/application` by itself doesn't solve the
  issue that `sprockets-rails` is no longer a dependency of `rails` and
  so must be added back to the Gemfile if it isn't a transitive
  dependency

This commit addresses those issues by ensuring sprockets-rails remains a
dependency in upgraded applications. The gem will not be added back for
application that previously skipped sprockets, which is determined by
the presence of `require "rails/all"` or `require "sprockets/railtie"`
in `config/application.rb`.

`active_storage:update` cannot be run inline since the bundle changes,
so the environment needs to fully reload.

Fixes #45295