Touch model records after ActiveStorage::Blob is analyzed
This fixes a race condition where a record can be requested and have a
cache entry built, before the initial `analyze_later` completes, which
will not be invalidated until something else updates the record.

1. Upload attachment (this will create an ActiveStorage::Blob,
   ActiveStorage::Attachment, and touch the model)
2. Enqueue analyze job
3. Request record
4. Build a cache for that record
5. Analyze the attachment.
6. Request the record again
7. Serve a cache that doesn't include data from the analyzers, because
   the cache was built before it got analyzed, and the updated_at
   hasn't changed.

This also invalidates cache entries when a blob is re-analyzed, which
is helpful if a bug is fixed in an analyzer or a new analyzer is added.

Fixes #45567