Correct/update the pluralization docs
----

`zero` is not a valid pluralization rule for most languages
(including `en`). Including `zero` in an `en` file (for example) would
cause the translation of that key to collide when translating into a
language that requires the `zero` key (e.g., Arabic)

```
en:
  car:
    one: I have %{count} car
    other: I have %{count} cars
```

Would translate into Arabic in this structure:

```
ar:
  car:
    zero: ...
    one: ...
    other: ...
```

The required `zero` key for Arabic would collide with the translation of
a `zero` key if it were to be provided in the `en` file.

CLDR defines explicit `0` and `1` keys for the cases where you need a
special string for the cases of `count == 0` or `count == 1`.

----

In this example, the `one` key should also include the `%{count}`
interpolation. A common mistake is to think that `one` is only for the
number 1. Instead, `one` is a category for any number that behaves
like 1.

Some languages use `one` for almost every number.
Some languages use `one` for numbers that end in "1" (like 1, 21, 151)
but that don't end in 11 (like 11, 111, 10311).
The same is true for the `zero`, and `two` categories. They are not
only used for 0 or 2, but any number that behaves like 0 or 2 in
that language.