Support transform block for class_attribute
This commit adds support for a transform block to `class_attribute`.  If
`class_attribute` is called with a block, the block will be applied to
an attribute value when that value is first read, and the result will be
memoized as the actual value of the attribute.

This provides an easy way to, for example, define an attribute that
calls `constantize` on its value, so that a constant reference may be
assigned before that constant has been defined / loaded.