Use already-loaded owner records when preloading
Prior to this commit, calling `includes` or `preload` on an
already-loaded relation or collection proxy would return a new relation
that ignored the loaded records.  This commit passes the loaded records
along to the new relation so that it can use them when preloading an
association.

Consider the following code:

  ```ruby
  posts.each do |post|
    # ...
  end

  # ...some intervening code...

  posts.includes(:comments).each do |post|
    post.comments.each do |comment|
      # ...
    end
  end
  ```

Before this commit, the code would execute 3 queries:

  ```
  SELECT "posts".* FROM "posts"
  SELECT "posts".* FROM "posts"
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" IN (?, ?, ?, ?, ?)  [["post_id", 1], ["post_id", 2], ["post_id", 3], ["post_id", 4], ["post_id", 5]]
  ```

After this commit, the code executes only 2 queries:

  ```
  SELECT "posts".* FROM "posts"
  SELECT "comments".* FROM "comments" WHERE "comments"."post_id" IN (?, ?, ?, ?, ?)  [["post_id", 1], ["post_id", 2], ["post_id", 3], ["post_id", 4], ["post_id", 5]]
  ```