Action View: use a local variable to hold the template buffer
This isn't quite working yet, there is a case failing with Builder.

But the idea is that if we no longer swap the buffer in `@output_buffer`
we can now reference it using the local variable instead of the instance
variable.

If we manage to make this work, it would open the door to further optimizations
such as Eruby's `chain_appends`.