Define #build_change_column_null_definition for MySQL and PostgreSQL
Defines API on the MySQL and PostgreSQL connection adapters for returning
schema definition objects that describe a column's NOT NULL constraint
being added or removed. This method should be called with the same arguments
as would be passed to #change_column_null.

Note that MySQL and PostgreSQL return different definition objects, due
to differences in their SQL needs. PG will return a ChangeColumnNullDefinition
object, whereas MySQL will return a ChangeColumnDefinition.