Fix argument error in nil pool config
If we don't have a `writing_pool_config` we shouldn't set it in the pool
manager, otherwise we'll get an error about trying to set a nil pool
config. I decided after this to delete the test for raising an argument
error when there's no `:writing` role in the pool. The reason for this
is that if you don't set `writing_role` in your application, you'll
still have a `:writing` role in your pool manager because that's Rails
default. Without eager loading, we only have the current role trying to
be set and may not have all the pools, so we can't really raise an error
assuming you didn't set up your roles correctly.

As a followup I'm going to deprecate `writing_role=` and
`reading_role=`. They aren't super useful and since Rails requires
either that the role is `:writing` or `writing_role` is set, but can't
error if it's not correctly set, I don't like the feature.

Fixes #45767