Update docs for `.unscoped` to be explicit about associations
We recently ran into an issue where we used `.unscoped` as part of a
method chain without realising that it would also unset the scoping of
the association. It feels like this would be useful information to
surface in the documentation so that there's a clear source of truth on
how the method behaves.