Remove incorrect 7.0 framework default
This was added in 60f7d49, which didn't quite make the cutoff for the
7.0 release, so it doesn't belong under `7.0`.

In addition, the value being set is the same as the underlying
configuration's default. This is likely an oversight because the
`load_default` value was changed during review from `true` to be
`Rails.env` dependent, and later to `false`.