Fix shared state from virtual attribute defaults
When an attribute without a database column defines a default proc, its
value is unknown until first accessed. Assignment of a new value
should not pass an unknown default as original attribute because
accessing it will memoize the proc value in both attribute instances
and cause them to share state.

Fixes #45817