Fix plugin generation inside applications
Previously, generating a plugin inside an application would create many
files that assume it should be able to run in isolation. The extraneous
generation also led to an error when using `rails generate plugin`
instead of `rails plugin new` where the dummy app created for tests
would try to use the same name as the parent application.

This commit fixes those issues by not generating the following when
inside a rails application:
- Gemfile
- `git init`
- a dummy application in test/

It also fixes the paths in the generated bin/rails and test_helper to
use the parent Rails application as a host.

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>