Speed up no-op nested attributes save
When we have two large collections, traversing one for each item in the
other has a complexity of O(n^2). This makes Active Record do a lot of
work in case when no record data has changed.

To speed it up, we create a map of existing recrods by record ID, and
use that inside the loop. This speeds up the no-op case (when no records
have changed) by 2x.