Reuse strict locals templates for any locals
This commit allows template lookups and render calls which find a
"strict local" template to reuse that template even when provided with
different locals.

This also makes Template#locals return nil for strict local templates,
to avoid leaking the details of the first lookup. Almost nothing calls
this method, so this should not be a significant change.