trying to get schema load working
stuck on the fact that the db tasks classes delegate to AR Base

what I want to do is pass the class in, so it's backwards compat but
were using regular *args. It's all no-doc so I think we should convert
it to **kwargs to make the inits cleaner.

I don't know if this is bad for other adapters though, might need to do
a deprecation cycle.