purge working, but ar metadata is a problem
ar metadata inhertis from base so we either need to treat it like we do
schema migrations or make sure that's still always on base.

note for later:
the only true way to isolate the rake task connections from Base in the
future would be to create a new connection handler like we do in the
tests. but i dont know how possible that is in the current iteration or
ever due to ar_metadata and other areas.